[Fix] [BE] WaggleLike 코드 수정
-

[Resolves] master
[Ref]
[Realted to]


Former-commit-id: 46e752f1b73df11a7bab47d8fac59016142514c3

+++ backend/Board/src/main/java/com/KL1verse/Waggle/repository/WaggleLikeRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface WaggleLikeRepository extends JpaRepository<WaggleLike, Long> {
 
-    Optional<WaggleLike> findByUserIdAndWaggleId_WaggleId(Long userId, Long waggleId);
+    Optional<WaggleLike> findByUserIdAndWaggleId_WaggleId(Integer userId, Long waggleId);
 
     @Query(value = "SELECT nickname, profile FROM user WHERE user_id = :userId", nativeQuery = true)
     List<Object[]> findNicknameAndProfileByUserId(Integer userId);
